Moved grocery APIs to backend and updated to module syntax. Removed bloat from README
